autoflex: wire up option functions, add IgnoreFieldNames option
This change introduces several changes to the AutoFlex options implementation to allow custom options to be passed to `flex.Expand` or `flex.Flatten`. The changes include:

- Adding a new `AutoFlexOptions` struct, which stores all configurable options
- Embedding the `AutoFlexOptions` struct into the `autoExpand` and `autoFlatten` flexers
- Adjusting the `AutoFlexOptionsFunc` signature to accept a pointer an `AutoFlexOptions` struct, allowing the default options to be adjusted

As a first use case, the `IgnoreFieldNames` option was added. This option is a slice of strings, with a default value of `[]string{"Tags"}`. This replaces the previously hardcoded logic which skipped processing of any fields named `Tags`, and instead makes this configurable. Test cases have been added to exercise the behavior with default options and overrides to remove defaults and add custom values.